+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+- structurizr-core: Adds name-value properties to dynamic view relationship views (https://github.com/structurizr/java/issues/316).
+- structurizr-dsl: Adds name-value properties to dynamic view relationship views.
 - structurizr-dsl: Fixes https://github.com/structurizr/java/issues/312 (!include doesn't work with files encoded as UTF-8 BOM).
 - structurizr-dsl: Adds a way to explicitly specify the order of relationships in dynamic views.